fixed the part for route saving, mapping to a graph
mapping to a graph previously only added edges between the new cell and
the cell where the drone is standing. however, this missed adding the
edges between the new cell and its surrounding cells, this was fixed.
now the targets are placed in the middle of the cell, so that the drone
moves more smoothly and get more range of view.